save,update were implemented. problem with delete and rest filter

+++ pay/src/main/java/furman/pay/repository/EmployeeRepository.java
@@ -4,11 +4,13 @@ import furman.pay.model.Employee;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import org.springframework.stereotype.Repository;
+
+import java.math.BigInteger;
 
 /**
  * Created by akoiro on 9/15/15.
  */
 @RepositoryRestResource(collectionResourceRel = "employee", path = "employee")
-public interface EmployeeRepository extends MongoRepository<Employee, Long>, QueryDslPredicateExecutor<Employee> {
+public interface EmployeeRepository extends MongoRepository<Employee, BigInteger>,
+        QueryDslPredicateExecutor<Employee> {
 }
